requirements.txt
azure-cognitiveservices-speech
PyAudio